+++ CHANGELOG.md
@@ -1,9 +1,12 @@
+# TBD
+### Added
+* Added ability to configure AWS Server Side Encryption for `S3S3Copier` via `copier-options.s3-server-side-encryption` configuration property.
+
 # 11.1.1 - 2018-02-15
 ### Fixed
 * Google FileSystem classes not being placed onto the mapreduce.application.classpath in S3MapReduceCp and DistCp mapreduce jobs
 ### Changed
 * Google FileSystem and S3 FileSystems added to mapreduce.application.classpath in circus-train-gcp and circus-train-aws respectively
-
 
 # 11.1.0 - 2018-02-05
 ### Fixed